Different approach to throwing away the old browser and docshell: Rather than creating a whole new tab, just throw away the browser (and subsequently, its docshell) and keep the tab object.
While this means we have to duplicate a bit of code from tabbrowser, it does simplify things under the hood: no TabClose/TabOpen/TabMove events are sent, and the original tab object prevails.

TODO: What to do when the selected tab is put on the bar tab? Also, this still doesnt' work right with TreeStyleTab :(