Get the tab title from the session history if available.
This title is available for all restored tabs already, which means we no longer have to query the nsINavHistoryService for that. What remains is a query to the nsIFaviconService for the favicon.

The only time we actually have to query the nsINavHistoryService is when we're faking a session history entry (opening a tab in the background).